meta.pm: fix logic for default meta values
The logic was wrong on a test during the handling of default
meta values.  If a page defined it's own author (say), the
code *should* have not bothered going through the motions of
using a default one.  However it was. The default one was
later being overridden with the page-specific one, so the
results were the same, albeit at the expense of running
redundant code.

A bit of whitespace clean-up at the same time.