remove.pm: add a remove action to the action list
Presently to remove a page via the web interface one must first select
'edit', then they are presented with 'remove'. This is counter-intuitive
and slow if you are trying to remove a lot of pages.

This patch adds a 'remove' action to the ikiwiki action-list which is
rendered on all pages (near e.g. 'Edit'). Page deletion is now two
clicks (remove, confirm) rather than three.