remove.pm: undef 'rebuild'
With displaying 'remove' in the action list, all pages need to be rebuilt for
it to show up. However, since it's not the only route to reach 'delete',
strictly a rebuild is not necessary (just a nice to have). As per the
documentation, 'undef' means "a rebuild could be needed in some circumstances,
but is not strictly required.".