Initial stab at a DuckDuckGo plugin
Derived from the Google plugin.

The form currently relies upon JavaScript to append a "site:" specification
to the search terms at submission time. I think this is unfortunate. An
alternative would be to route all searches through the IkiWiki CGI, but I
felt that would be unnecessarily slow.

DuckDuckGo themselves provide an iframe-based solution for embedding a search
box into a site; I felt that this would not be attractive to privacy-focussed
IkiWiki users. Their iframe approach also uses JavaScript to append the "site:"
specification.