Enable dark mode for the anti-theme
In order for a user's browser or OS-preference vis-a-vis Dark Mode to be
honoured in a theme, the top-level HTML element needs to have 'dark' set
for its color-scheme property. Specifying "light dark" means that both
are supported.

This patch adds the minimum necessary for the default anti-theme to
report that Dark mode is supported, and thus to be selected according
to user preference. The actual colours used are whatever the Browser
chooses, just like with light/default mode.

This does not touch the separate themes, which are unaffected. A
downstream user who has customized look & feel via local.css may see
different results after this patch, especially if their style choices
did not consider dark or light mode preferences.

Thanks Dagfinn Ilmari Mannsåker for the inspiration.