uncompress our logo so we get diffs
Before: 6.2K, after: 32K, so we're adding 28K to our git history,
permanently. But then this is compressed by git and now we get
meaningful diffs: i think in the long run it's a huge win.

It should also make it readable in a web browser: Firefox currently
chokes on the .svgz.